Fix dependencies between modules, remove unused dependencies, separate dagger and anvil modules
