+++ api/anime-details/src/main/java/com/majo/anime_details/di/AnimeDetailsApiModule.kt
@@ -1,7 +1,7 @@
 package com.majo.anime_details.di
 
 import com.majo.anime_details.AnimeDetailsApi
-import com.majo.shikimori.dagger.anvil.AppScope
+import com.majo.shikimori.anvil.AppScope
 import com.majo.shikimori.retrofit.RetrofitFactory
 import com.squareup.anvil.annotations.ContributesTo
 import dagger.Module
