+++ api/manga-list/src/main/java/com/majo/shikimori/manga_list/di/MangaListApiModule.kt
@@ -1,6 +1,6 @@
 package com.majo.shikimori.manga_list.di
 
-import com.majo.shikimori.dagger.anvil.AppScope
+import com.majo.shikimori.anvil.AppScope
 import com.majo.shikimori.manga_list.MangaListApi
 import com.majo.shikimori.retrofit.RetrofitFactory
 import com.squareup.anvil.annotations.ContributesTo
