+++ build.gradle
@@ -5,4 +5,5 @@ plugins {
     id 'org.jetbrains.kotlin.android' version '1.7.20' apply false
     id 'com.squareup.anvil' version "2.4.3"
     id 'org.jetbrains.kotlin.plugin.serialization' version '1.7.20'
+    id 'com.savvasdalkitsis.module-dependency-graph' version '0.10'
 }
