+++ common/anvil/build.gradle.kts
@@ -1,0 +1,4 @@
+plugins {
+    id ("convention.android-library")
+    id("convention.dagger")
+}
