+++ common/anvil/src/main/java/com/majo/shikimori/anvil/AppScope.kt
@@ -1,3 +1,3 @@
-package com.majo.shikimori.dagger.anvil
+package com.majo.shikimori.anvil
 
 abstract class AppScope private constructor()
