+++ common/anvil/src/main/java/com/majo/shikimori/anvil/ScreenScope.kt
@@ -1,3 +1,3 @@
-package com.majo.shikimori.dagger.anvil
+package com.majo.shikimori.anvil
 
 abstract class ScreenScope private constructor()
