+++ common/dagger/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id ("convention.android-library")
+    id("convention.android-library")
     id("convention.dagger")
     id("convention.compose")
 }
