+++ common/dagger/src/main/java/com/majo/shikimori/dagger/ComposableDagger.kt
@@ -4,6 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 
+@Suppress("UNCHECKED_CAST")
 @Composable
 inline fun <reified T : ViewModel> daggerViewModel(
     key: String? = null,
