--- common/dagger/src/main/java/com/majo/shikimori/dagger/PerActivity.kt
@@ -1,7 +1,0 @@
-package com.majo.shikimori.dagger
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-annotation class PerActivity
