--- common/dagger/src/main/java/com/majo/shikimori/dagger/anvil/ActivityScope.kt
@@ -1,3 +1,0 @@
-package com.majo.shikimori.dagger.anvil
-
-abstract class ActivityScope private constructor()
