--- common/dagger/src/main/java/com/majo/shikimori/dagger/anvil/FragmentScope.kt
@@ -1,3 +1,0 @@
-package com.majo.shikimori.dagger.anvil
-
-abstract class FragmentScope private constructor()
