+++ common/navigation/build.gradle.kts
@@ -2,7 +2,3 @@ plugins {
     id("convention.android-library")
     id("convention.compose")
 }
-
-dependencies {
-    implementation(projects.common.android)
-}
