+++ common/retrofit/src/main/java/com/majo/shikimori/retrofit/RetrofitFactory.kt
@@ -1,6 +1,6 @@
 package com.majo.shikimori.retrofit
 
-import com.majo.shikimori.dagger.anvil.AppScope
+import com.majo.shikimori.anvil.AppScope
 import com.squareup.anvil.annotations.ContributesBinding
 import retrofit2.Retrofit
 import javax.inject.Inject
