+++ features/anime-details-impl/build.gradle.kts
@@ -7,6 +7,7 @@ plugins {
 dependencies {
     implementation(projects.api.animeDetails)
     implementation(projects.common.dagger)
+    implementation(projects.common.anvil)
     implementation(projects.common.android)
     implementation(projects.common.navigation)
     implementation(projects.common.designSystem)
