+++ features/anime-details-impl/src/main/java/com/majo/shikimori/anime_details/di/AnimeDetailsComponent.kt
@@ -1,15 +1,14 @@
 package com.majo.shikimori.anime_details.di
 
 import com.majo.shikimori.anime_details.AnimeDetailsViewModel
+import com.majo.shikimori.anvil.ScreenScope
 import com.majo.shikimori.dagger.PerScreen
-import com.majo.shikimori.dagger.anvil.ScreenScope
 import com.squareup.anvil.annotations.MergeComponent
 import dagger.BindsInstance
 import dagger.Component
 
 @MergeComponent(
     scope = ScreenScope::class,
-    modules = [AnimeDetailsModule::class],
     dependencies = [AnimeDetailsDependencies::class],
 )
 @PerScreen
