+++ features/anime-details-impl/src/main/java/com/majo/shikimori/anime_details/di/AnimeDetailsDependencies.kt
@@ -2,8 +2,8 @@ package com.majo.shikimori.anime_details.di
 
 import com.majo.anime_details.AnimeDetailsApi
 import com.majo.shikimori.android.ErrorConverter
+import com.majo.shikimori.anvil.AppScope
 import com.majo.shikimori.dagger.ComponentDependencies
-import com.majo.shikimori.dagger.anvil.AppScope
 import com.squareup.anvil.annotations.ContributesTo
 
 @ContributesTo(AppScope::class)
