+++ features/anime-details-impl/src/main/java/com/majo/shikimori/anime_details/di/AnimeDetailsModule.kt
@@ -2,11 +2,15 @@ package com.majo.shikimori.anime_details.di
 
 import com.majo.shikimori.anime_details.AnimeDetailsViewModel
 import com.majo.shikimori.anime_details.mvi.AnimeDetailsFeatureBuilder
+import com.majo.shikimori.anvil.AppScope
+import com.majo.shikimori.anvil.ScreenScope
 import com.majo.shikimori.dagger.PerScreen
+import com.squareup.anvil.annotations.ContributesTo
 import dagger.Module
 import dagger.Provides
 
 @Module
+@ContributesTo(ScreenScope::class)
 class AnimeDetailsModule {
 
     @Provides
