+++ features/anime-details-impl/src/main/java/com/majo/shikimori/anime_details/di/AnimeDetailsScreenProviderModule.kt
@@ -2,7 +2,7 @@ package com.majo.shikimori.anime_details.di
 
 import com.majo.shikimori.anime_details.AnimeDetailsScreenProvider
 import com.majo.shikimori.anime_details.AnimeDetailsScreenProviderImpl
-import com.majo.shikimori.dagger.anvil.AppScope
+import com.majo.shikimori.anvil.AppScope
 import com.squareup.anvil.annotations.ContributesTo
 import dagger.Module
 import dagger.Provides
