+++ features/animelist-impl/build.gradle
@@ -9,6 +9,7 @@ dependencies {
     implementation(projects.features.animelistPublic)
     implementation(projects.features.animeDetailsPublic)
     implementation(projects.common.dagger)
+    implementation(projects.common.anvil)
     implementation(projects.common.android)
     implementation(projects.common.navigation)
     implementation(projects.common.mvi)
