+++ features/animelist-impl/src/main/java/com/majo/shikimori/animelist/AnimeListScreenProviderImpl.kt
@@ -5,10 +5,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import com.majo.shikimori.anime_details.AnimeDetailsScreenProvider
 import com.majo.shikimori.animelist_public.AnimeListScreenProvider
-import com.majo.shikimori.dagger.PerScreen
-import com.majo.shikimori.dagger.anvil.ScreenScope
 import com.majo.shikimori.navigation.NavigationItem
-import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
 
 class AnimeListScreenProviderImpl @Inject constructor(
