+++ features/animelist-impl/src/main/java/com/majo/shikimori/animelist/di/AnimeListDependencies.kt
@@ -3,8 +3,8 @@ package com.majo.shikimori.animelist.di
 import com.majo.shikimori.android.ErrorConverter
 import com.majo.shikimori.anime_details.AnimeDetailsScreenProvider
 import com.majo.shikimori.animelist.AnimeListApi
+import com.majo.shikimori.anvil.AppScope
 import com.majo.shikimori.dagger.ComponentDependencies
-import com.majo.shikimori.dagger.anvil.AppScope
 import com.squareup.anvil.annotations.ContributesTo
 
 @ContributesTo(AppScope::class)
