+++ features/animelist-impl/src/main/java/com/majo/shikimori/animelist/di/AnimeListModule.kt
@@ -1,14 +1,15 @@
 package com.majo.shikimori.animelist.di
 
-import com.majo.shikimori.animelist.AnimeListScreenProviderImpl
 import com.majo.shikimori.animelist.AnimeListViewModel
 import com.majo.shikimori.animelist.mvi.AnimeListFeatureBuilder
-import com.majo.shikimori.animelist_public.AnimeListScreenProvider
+import com.majo.shikimori.anvil.ScreenScope
 import com.majo.shikimori.dagger.PerScreen
+import com.squareup.anvil.annotations.ContributesTo
 import dagger.Module
 import dagger.Provides
 
 @Module
+@ContributesTo(ScreenScope::class)
 class AnimeListModule {
 
     @Provides
