+++ features/animelist-impl/src/main/java/com/majo/shikimori/animelist/di/AnimeListScreenProviderModule.kt
@@ -3,7 +3,7 @@ package com.majo.shikimori.animelist.di
 import com.majo.shikimori.anime_details.AnimeDetailsScreenProvider
 import com.majo.shikimori.animelist.AnimeListScreenProviderImpl
 import com.majo.shikimori.animelist_public.AnimeListScreenProvider
-import com.majo.shikimori.dagger.anvil.AppScope
+import com.majo.shikimori.anvil.AppScope
 import com.squareup.anvil.annotations.ContributesTo
 import dagger.Module
 import dagger.Provides
