+++ features/home/src/main/java/com/majo/shikimori/home/di/HomeListComponent.kt
@@ -1,8 +1,8 @@
 package com.majo.shikimori.home.di
 
 import com.majo.shikimori.animelist_public.AnimeListScreenProvider
+import com.majo.shikimori.anvil.ScreenScope
 import com.majo.shikimori.dagger.PerScreen
-import com.majo.shikimori.dagger.anvil.ScreenScope
 import com.majo.shikimori.manga_list.MangaListScreenProvider
 import com.squareup.anvil.annotations.MergeComponent
 import dagger.Component
