+++ features/manga-list-impl/src/main/java/com/majo/shikimori/manga_list/MangaListInteractor.kt
@@ -1,8 +1,8 @@
 package com.majo.shikimori.manga_list
 
 import com.majo.shikimori.android.ErrorConverter
+import com.majo.shikimori.anvil.ScreenScope
 import com.majo.shikimori.dagger.PerScreen
-import com.majo.shikimori.dagger.anvil.ScreenScope
 import com.majo.shikimori.manga_list.mvi.MangaListActor.Companion.INIT_PAGE
 import com.majo.shikimori.manga_list.mvi.entity.MangaListInternalAction
 import com.squareup.anvil.annotations.ContributesBinding
