+++ features/manga-list-impl/src/main/java/com/majo/shikimori/manga_list/di/MangaListDependencies.kt
@@ -1,9 +1,8 @@
 package com.majo.shikimori.manga_list.di
 
-import android.content.Context
 import com.majo.shikimori.android.ErrorConverter
+import com.majo.shikimori.anvil.AppScope
 import com.majo.shikimori.dagger.ComponentDependencies
-import com.majo.shikimori.dagger.anvil.AppScope
 import com.majo.shikimori.manga_list.MangaListApi
 import com.squareup.anvil.annotations.ContributesTo
