+++ features/manga-list-impl/src/main/java/com/majo/shikimori/manga_list/di/MangaListScreenProviderModule.kt
@@ -1,6 +1,6 @@
 package com.majo.shikimori.manga_list.di
 
-import com.majo.shikimori.dagger.anvil.AppScope
+import com.majo.shikimori.anvil.AppScope
 import com.majo.shikimori.manga_list.MangaListScreenProvider
 import com.majo.shikimori.manga_list.MangaListScreenProviderImpl
 import com.squareup.anvil.annotations.ContributesTo
