+++ features/manga-list-public/build.gradle.kts
@@ -1,10 +1,7 @@
 plugins {
    id("convention.android-library")
-    id("convention.dagger")
-    id("convention.compose")
 }
 
 dependencies {
-    implementation(projects.common.dagger)
-    implementation(projects.common.android)
+   implementation(libs.compose.navigation)
 }
