+++ gradle/libs.versions.toml
@@ -12,7 +12,6 @@
     dagger = "2.43.2"
     retrofit = "2.9.0"
     coroutines = "1.3.9"
-    glide = "4.13.2"
 	kotlin = "1.6.21"
 	buildGradle = "7.3.1"
 	accompanist = "0.28.0"
@@ -45,7 +44,6 @@
 	compose-navigation = { group = "androidx.navigation", name = "navigation-compose", version.ref = "compose-navigation" }
 	coreKtx = { group = "androidx.core", name = "core-ktx", version.ref = "core" }
 	accompanistSystemuicontroller = { group = "com.google.accompanist", name = "accompanist-systemuicontroller", version.ref = "accompanist"}
-	lifecycleLivedataKtx = { group = "androidx.lifecycle", name = "lifecycle-livedata-ktx", version.ref = "lifecycle" }
 	lifecycleRuntimeCompose = { group = "androidx.lifecycle", name = "lifecycle-runtime-compose", version.ref = "lifecycle" }
 	lifecycleViewModel = { group = "androidx.lifecycle", name = "lifecycle-viewmodel-ktx", version.ref = "lifecycle" }
 	lifecycleViewModelCompose = { group = "androidx.lifecycle", name = "lifecycle-viewmodel-compose", version.ref = "lifecycle" }
