+++ settings.gradle.kts
@@ -37,3 +37,4 @@ include(":features:anime-details-public")
 include(":api:manga-list")
 include(":features:manga-list-public")
 include(":features:manga-list-impl")
+include(":common:anvil")
