+++ shikimori-app/src/main/java/com/majo/shikimori/di/component/AppComponent.kt
@@ -1,8 +1,8 @@
 package com.majo.shikimori.di.component
 
 import com.majo.shikimori.App
+import com.majo.shikimori.anvil.AppScope
 import com.majo.shikimori.dagger.AppComponentDependencies
-import com.majo.shikimori.dagger.anvil.AppScope
 import com.squareup.anvil.annotations.MergeComponent
 import dagger.BindsInstance
 import dagger.Component
