+++ shikimori-app/src/main/java/com/majo/shikimori/di/module/AppModule.kt
@@ -3,7 +3,7 @@ package com.majo.shikimori.di.module
 import android.content.Context
 import android.content.res.Resources
 import com.majo.shikimori.App
-import com.majo.shikimori.dagger.anvil.AppScope
+import com.majo.shikimori.anvil.AppScope
 import com.squareup.anvil.annotations.ContributesTo
 import dagger.Module
 import dagger.Provides
